share user group modals across states
