+++ src/Components/UserGroups/CreateGroupModal.tsx
@@ -1,5 +1,3 @@
-import { useState, useContext } from 'react';
-import { createUserGroup } from '../../API/userCalls';
 import GenericGroupModal from './GenericGroupModal';
 import './CreateGroupModal.css';
 
@@ -20,12 +18,6 @@ export const CreateGroupModal: React.FC<CreateGroupModalProps> = ({
   onSuccess,
   initialValues
 }) => {
-  const handleSubmit = async (values: { name: string; users: string[]; description?: string }) => {
-    // Call the API to create a new group
-    const newGroup = await createUserGroup(values);
-    return newGroup;
-  };
-
   return (
     <GenericGroupModal
       visible={visible}
@@ -34,9 +26,8 @@ export const CreateGroupModal: React.FC<CreateGroupModalProps> = ({
       title="Add Collaborator"
       initialValues={initialValues}
       submitButtonText="Save Group & Send Invites"
-      onSubmit={handleSubmit}
     />
   );
 };
 
-export default CreateGroupModal; 
+export default CreateGroupModal;
