+++ src/Components/UserGroups/EditGroupModal.tsx
@@ -1,7 +1,4 @@
-import { useState, useContext } from 'react';
-import { message } from 'antd';
-import { updateUserGroup, UserGroupResponseDataType } from '../../API/userCalls';
-import { UserGroupDataType, UserDataType } from '../../Types';
+import { UserGroupDataType } from '../../Types';
 import GenericGroupModal from './GenericGroupModal';
 
 interface EditGroupModalProps {
@@ -17,27 +14,6 @@ export const EditGroupModal: React.FC<EditGroupModalProps> = ({
   onSuccess,
   group
 }) => {
-  const handleSubmit = async (values: { name: string; users: string[]; description?: string }, groupId?: number) => {
-    if (!groupId) throw new Error('Group ID is required for updating a group');
-
-    // Create updated group data, preserving existing fields except users
-    const { users, ...groupWithoutUsers } = group;
-    
-    // Convert to UserGroupResponseDataType which is what the API expects
-    const updatedGroup: UserGroupResponseDataType = {
-      id: groupId,
-      name: values.name,
-      users: values.users,
-      user_ids: group.user_ids,
-      signal_ids: group.signal_ids,
-      collaborator_map: group.collaborator_map
-    };
-
-    // Call the API to update the group
-    const result = await updateUserGroup(groupId, updatedGroup);
-    return result;
-  };
-
   return (
     <GenericGroupModal
       visible={visible}
@@ -49,9 +25,8 @@ export const EditGroupModal: React.FC<EditGroupModalProps> = ({
         users: group?.users ? group.users.map(u => u.email) : [],
         description: group?.description || ''
       }}
-      groupId={group?.id}
+      group={group}
       submitButtonText="Update Group"
-      onSubmit={handleSubmit}
     />
   );
 };
