+++ src/Components/UserGroups/index.tsx
@@ -1,5 +1,4 @@
 export * from './UserGroupsList';
-export * from './UserGroupForm';
 export * from './CreateGroupModal';
 export * from './EditGroupModal';
 export * from './GenericGroupModal';
